requirements.txt
# === Core API Framework ===
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.5

# === Machine Learning / CLIP Classification ===
torch==2.2.2
torchvision==0.17.2
transformers==4.57.0
pillow==10.2.0
numpy==1.26.4

# === Face Verification ===
facenet-pytorch==2.6.0

# === OCR (ID text extraction) ===
easyocr==1.7.2
opencv-python-headless==4.11.0.86
scikit-image==0.25.2
scipy==1.15.3
Shapely==2.1.2
pyclipper==1.3.0.post6
python-bidi==0.6.7
